[Refactor] useNextScreenNavigation hook 분리 (#15)
